requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
chromadb>=0.4.18
sentence-transformers>=2.2.2
watchdog>=3.0.0
python-dotenv>=1.0.0
toml>=0.10.2
pyyaml>=6.0
requests>=2.31.0
